initiate egress unpacks executor outputs ready for egress
